Tests can now be called from and supply data to matlab -- still working on the protocol
Tests_run now controls the comm port and listens for incoming test requests and supplies the results.

Next I need to figure out what is wrong with EEPROM -- it was working 2 weeks ago, but the readback is failing now. Likely culprit is GPIO_init -- need to write pin and clock inits for spi and EEPROM anyway.